% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AsCas12a}
\alias{AsCas12a}
\title{AsCas12a CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(AsCas12a, package="crisprBase")
}
\description{
CrisprNuclease object for the Wildtype Acidaminococcus Cas12a (AsCas12a)
    nuclease.
}
\details{
The AsCas12a nuclease recognizes TTTV PAM sequences. Spacer
    sequences must be located downstream of PAM sequences.
}
\keyword{datasets}
