% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BE4max}
\alias{BE4max}
\title{BE4max BaseEditor object}
\format{
BaseEditor object.
}
\usage{
data(BE4max, package="crisprBase")
}
\description{
BaseEditor for the cytosine base editor CRISPR/Cas9 system
    BE4max. Editing weights were obtained from
    https://doi.org/10.1016/j.cell.2020.05.037
}
\details{
BaseEditor for the cytosine base editor
    CRISPR/Cas9 system BE4max. Editing weights were obtained from
    \url{https://doi.org/10.1016/j.cell.2020.05.037}.
}
\keyword{datasets}
