% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CasRx}
\alias{CasRx}
\title{CasRx CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(CasRx, package="crisprBase")
}
\description{
CrisprNuclease object for the Cas13d-NLS from Ruminococcus
    flavefaciens strain XPD3002 nuclease (RNase).
}
\details{
The CasRx nuclease was derived from Cas13d Ruminococcus
    flavefaciens string XPD3002. See \url{10.1016/j.cell.2018.02.033}.
}
\keyword{datasets}
