% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/CrisprNucleaseBaseEditor-class.R
\docType{class}
\name{baseEditorName}
\alias{baseEditorName}
\alias{baseEditorName<-}
\alias{editingWeights}
\alias{editingWeights<-}
\alias{CrisprNucleaseBaseEditor-class}
\alias{baseEditorName,CrisprNucleaseBaseEditor-method}
\alias{baseEditorName<-,CrisprNucleaseBaseEditor-method}
\alias{editingWeights,CrisprNucleaseBaseEditor-method}
\alias{editingWeights<-,CrisprNucleaseBaseEditor-method}
\title{An S4 class to represent a CRISPR nuclease with base editor.}
\usage{
baseEditorName(object)

baseEditorName(object) <- value

editingWeights(object)

editingWeights(object) <- value

\S4method{baseEditorName}{CrisprNucleaseBaseEditor}(object)

\S4method{baseEditorName}{CrisprNucleaseBaseEditor}(object) <- value

\S4method{editingWeights}{CrisprNucleaseBaseEditor}(object)

\S4method{editingWeights}{CrisprNucleaseBaseEditor}(object) <- value
}
\arguments{
\item{object}{\linkS4class{CrisprNucleaseBaseEditor} object.}

\item{value}{Value to replaced with.}
}
\value{
A CrisprNucleaseWithBaseEdtitor object
}
\description{
An S4 class to represent a CRISPR nuclease with base editor.
}
\section{Slots}{

\describe{
\item{\code{baseEditorName}}{Name of the base editor enzyme.}

\item{\code{editingWeights}}{Matrix of editing weights.}
}}

\section{Constructors}{

    Use the constructor \code{link{CrisprNucleaseBaseEditor}} to create
    a CrisprNucleaseBaseEditor object.
}

