% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Csm}
\alias{Csm}
\title{Csm CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(Csm, package="crisprBase")
}
\description{
CrisprNuclease object for the RNA-targeting
    Csm complex from Streptococcus thermophilus
}
\details{
The specific Csm complex is an RNA-targeting nuclease derived from 
    Streptococcus thermophilus. There is no preferred PAM sequences,
    and the default (optimal) spacer length is 32nt. 
    See \url{https://doi.org/10.1038/s41587-022-01649-9}.
}
\keyword{datasets}
