% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{enAsCas12a}
\alias{enAsCas12a}
\title{enAsCas12a CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(enAsCas12a, package="crisprBase")
}
\description{
CrisprNuclease object for the Enhanced Acidaminococcus Cas12a (AsCas12a)
    nuclease.
}
\details{
The enAsCas12a nuclease recognizes an extended set of PAM sequences
    beyong the canonical TTTV sequence for AsCas12a. Spacer sequences must 
    be located downstream of PAM sequences.
}
\keyword{datasets}
