% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{extractPamFromProtospacer}
\alias{extractPamFromProtospacer}
\title{Extract PAM sequences from protospacer sequences}
\usage{
extractPamFromProtospacer(protospacers, object)
}
\arguments{
\item{protospacers}{Character vector of protospacer sequences.}

\item{object}{\code{CrisprNuclease} corresponding to the protospacer 
sequences.}
}
\value{
Character vector of PAM sequences of length equal to that of the 
    \code{protospacers} character vector.
}
\description{
Extract PAM sequences from protospacer sequences using
    information stored in a CrisprNuclease object.
}
\examples{
data(SpCas9, AsCas12a, package="crisprBase")
# Extracting PAM sequences from Cas9 protospacers:
protospacers <- c("AGGTGCTGATTGTAGTGCTGCGG",
                  "AGGTGCTGATTGTAGTGCTGAGG")
extractPamFromProtospacer(protospacers, SpCas9)
# Extracting PAM sequences from Cas12a protospacers:
protospacers <- c("TTTAAGGTGCTGATTGTAGTGCTGTGT",
                  "TTTCAGGTGCTGATTGTAGTGCTGAAA")
extractPamFromProtospacer(protospacers, AsCas12a)

}
\author{
Jean-Philippe Fortin
}
