% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{extractSpacerFromProtospacer}
\alias{extractSpacerFromProtospacer}
\title{Extract spacer sequences from protospacer sequences}
\usage{
extractSpacerFromProtospacer(protospacers, object)
}
\arguments{
\item{protospacers}{Character vector of protospacer sequences.}

\item{object}{\code{CrisprNuclease} corresponding to the protospacer 
sequences.}
}
\value{
Character vector of spacer sequences of length equal to that of the 
    \code{protospacers} character vector.
}
\description{
Extract spacer sequences from protospacer sequences using
    information stored in a CrisprNuclease object.
}
\examples{
data(SpCas9, AsCas12a, package="crisprBase")
# Extracting spacer sequences from Cas9 protospacers:
protospacers <- c("AGGTGCTGATTGTAGTGCTGCGG",
                  "AGGTGCTGATTGTAGTGCTGAGG")
extractSpacerFromProtospacer(protospacers, SpCas9)
# Extracting spacer sequences from Cas12a protospacers:
protospacers <- c("TTTAAGGTGCTGATTGTAGTGCTGTGT",
                  "TTTCAGGTGCTGATTGTAGTGCTGAAA")
extractSpacerFromProtospacer(protospacers, AsCas12a)

}
\author{
Jean-Philippe Fortin
}
