% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAvailableCrisprNucleases}
\alias{getAvailableCrisprNucleases}
\title{Return list of available CrisprNuclease objects in crisprBase}
\usage{
getAvailableCrisprNucleases()
}
\value{
Character vector of available \code{CrisprNuclease} objects found
    in \code{crisprBase}.
}
\description{
Return list of available CrisprNuclease objects in crisprBase.
}
\examples{
getAvailableCrisprNucleases()

}
\author{
Jean-Philippe Fortin
}
