% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{getCutSiteRanges}
\alias{getCutSiteRanges}
\title{Construct a cut site GRanges from a list of PAM sites}
\usage{
getCutSiteRanges(
  gr = NULL,
  seqnames = NULL,
  pam_site = NULL,
  strand = NULL,
  nuclease = NULL
)
}
\arguments{
\item{gr}{GRanges object of width 1 specifying the coordinates
of the first nucleotide of the PAM sequences.}

\item{seqnames}{Character vector of genomic sequence names.
Ignored if \code{gr} is not NULL.}

\item{pam_site}{Numeric vector specifying the coordinates of the
first nucleotide of the PAM sequences corresponding to the
PAM sequences. Ignored if \code{gr} is not NULL.}

\item{strand}{Character vector specifying the strand of the PAM. 
Ignored if \code{gr} is not NULL.}

\item{nuclease}{CrisprNuclease object.}
}
\value{
GRanges object representing genomic coordinates of the cut sites.
}
\description{
Construct a cut site GRanges from a list of PAM sites 
    using information stored in a CrisprNuclease object.
}
\examples{
data(SpCas9, AsCas12a, package="crisprBase")
library(GenomicRanges)
gr <- GRanges("chr10",
              IRanges(start=c(100,120), width=1),
              strand=c("+","-"))
getCutSiteRanges(gr, nuclease=SpCas9)
getCutSiteRanges(gr, nuclease=AsCas12a)

}
\author{
Jean-Philippe Fortin
}
