% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MAD7}
\alias{MAD7}
\title{MAD7 CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(MAD7, package="crisprBase")
}
\description{
CrisprNuclease object for the MAD7 nuclease (Cas12a-like nuclease)
}
\details{
The MAD7 nuclease recognizes YTTV PAM sequences. Spacer
    sequences must be located downstream of PAM sequences.
}
\keyword{datasets}
