% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{restrictionEnzymes}
\alias{restrictionEnzymes}
\title{List of Nuclease objects representing common restriction enzymes}
\format{
List of Nuclease objects.
}
\usage{
data(restrictionEnzymes, package="crisprBase")
}
\description{
List of Nuclease objects representing common restriction enzymes
    from REBASE database.
}
\details{
List of Nuclease objects representing common restriction enzymes
    from REBASE database.
}
\keyword{datasets}
