% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SaCas9}
\alias{SaCas9}
\title{SaCas9 CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(SaCas9, package="crisprBase")
}
\description{
CrisprNuclease object for the wildtype Staphylococcus aureus Cas9 (SaCas9)
    nuclease.
}
\details{
The AsCas9 nuclease recognizes NNGRRT PAM sequences. Spacer
    sequences must be located upstream of PAM sequences. Editing weights 
    were obtained from \url{doi:10.1038/nature14299}.
}
\keyword{datasets}
