% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpCas9}
\alias{SpCas9}
\title{SpCas9 CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(SpCas9, package="crisprBase")
}
\description{
CrisprNuclease object for the wildtype Streptococcus pyogenes Cas9 (SpCas9)
    nuclease.
}
\details{
The SpCas9 nuclease recognizes NGG PAM sequences. Spacer
    sequences must be located upstream of PAM sequences.
}
\keyword{datasets}
