% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpG}
\alias{SpG}
\title{SpG CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(SpG, package="crisprBase")
}
\description{
CrisprNuclease object for the engineered Streptococcus pyogenes Cas9
    SpG nuclease.
}
\details{
The SpG nuclease recognizes NGN PAM sequences. Spacer
    sequences must be located upstream of PAM sequences. Editing weights 
    were obtained from \url{doi:10.1038/s41467-022-28884-7}.
}
\keyword{datasets}
