% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpGCas9}
\alias{SpGCas9}
\title{SpGCas9 CrisprNuclease object}
\format{
CrisprNuclease object.
}
\usage{
data(SpGCas9, package="crisprBase")
}
\description{
CrisprNuclease object for the engineered Streptococcus pyogenes Cas9
    SpG nuclease.
}
\details{
The SpGCas9 nuclease recognizes NGN PAM sequences. Spacer
    sequences must be located upstream of PAM sequences.
}
\keyword{datasets}
