% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R,
%   R/CrisprNucleaseBaseEditor-class.R
\docType{class}
\name{baseEditorName}
\alias{baseEditorName}
\alias{baseEditorName<-}
\alias{editingWeights}
\alias{editingWeights<-}
\alias{editingStrand}
\alias{editingStrand<-}
\alias{CrisprNucleaseBaseEditor-class}
\alias{CrisprNucleaseBaseEditor}
\alias{show,CrisprNucleaseBaseEditor-method}
\alias{baseEditorName,CrisprNucleaseBaseEditor-method}
\alias{baseEditorName<-,CrisprNucleaseBaseEditor-method}
\alias{editingWeights,CrisprNucleaseBaseEditor-method}
\alias{editingWeights<-,CrisprNucleaseBaseEditor-method}
\alias{editingStrand,CrisprNucleaseBaseEditor-method}
\alias{editingStrand<-,CrisprNucleaseBaseEditor-method}
\title{An S4 class to represent a CRISPR nuclease with base editing}
\usage{
baseEditorName(object)

baseEditorName(object) <- value

editingWeights(object)

editingWeights(object) <- value

editingStrand(object)

editingStrand(object) <- value

CrisprNucleaseBaseEditor(
  CrisprNuclease,
  baseEditorName = NA_character_,
  editingStrand = c("original", "opposite"),
  editingWeights = NULL
)

\S4method{show}{CrisprNucleaseBaseEditor}(object)

\S4method{baseEditorName}{CrisprNucleaseBaseEditor}(object)

\S4method{baseEditorName}{CrisprNucleaseBaseEditor}(object) <- value

\S4method{editingWeights}{CrisprNucleaseBaseEditor}(object)

\S4method{editingWeights}{CrisprNucleaseBaseEditor}(object) <- value

\S4method{editingStrand}{CrisprNucleaseBaseEditor}(object)

\S4method{editingStrand}{CrisprNucleaseBaseEditor}(object) <- value
}
\arguments{
\item{object}{\linkS4class{CrisprNucleaseBaseEditor} object.}

\item{value}{Value to replaced with.}

\item{CrisprNuclease}{A \linkS4class{CrisprNuclease} object.}

\item{baseEditorName}{String specifying base editor name.}

\item{editingStrand}{String indicating which strand with
respect to the target protospacer sequence will be 
edited. Must be either "original" or "opposite".
"original" by default.}

\item{editingWeights}{Numeric matrix of editing weights.
Column names must be indicating relative position to the PAM site.
Row names must be of the form "X2Y" where "X" represents the origin
base, and "Y" represents the subtituted base. For instance, "C2T"
indicates the row corresponding to C to T editing.}
}
\value{
A CrisprNucleaseWithBaseEditor object
}
\description{
An S4 class to represent a CRISPR nuclease with base editing
}
\section{Functions}{
\itemize{
\item \code{CrisprNucleaseBaseEditor}: Create a \linkS4class{CrisprNucleaseBaseEditor} object
}}

\section{Slots}{

\describe{
\item{\code{baseEditorName}}{Name of the base editor enzyme.}

\item{\code{editingWeights}}{Matrix of editing weights.}

\item{\code{editingStrand}}{String indicating which strand with
respect to the target protospacer sequence will be 
edited. Must be either "original" or "opposite".
"original" by default.}
}}

\section{Constructors}{

    Use the constructor \code{link{CrisprNucleaseBaseEditor}} to create
    a CrisprNucleaseBaseEditor object.
}

\section{Accessors}{

\describe{
    \item{\code{baseEditorName}:}{To get the name of the base editor.}
    \item{\code{editingWeights}:}{To return the matrix of editing weights.}
    \item{\code{editingStrand}:}{To return the editing strand.}
}
}

\section{Setters}{

\describe{
    \item{\code{baseEditorName<-}:}{To change the name of the base editor.}
    \item{\code{editingWeights<-}:}{To change the matrix of editing weights.}
    \item{\code{editingStrand<-}:}{To change the editing strand.}
}
}

\examples{
# Creating an object for BE4max (C to T editor)
# based on experimental weights

ws <- c(0.7, 0.7, 0.8, 1.8, 1, 2, 1.4, 1.2, 2.3, 1.3, 2.4, 2.2, 3.4, 
      2.2, 2.1, 3.5, 5.8, 16.2, 31.8, 63.2, 90.3, 100, 87, 62, 31.4, 
      16.3, 10, 5.6, 3.3, 1.9, 1.8, 2.4, 1.7, 0.5, 0.2, 0.1)
ws <- matrix(ws, nrow=1, ncol=length(ws))
rownames(ws) <- "C2T"
colnames(ws) <- -36:-1
data(SpCas9, package="crisprBase")
BE4max <- CrisprNucleaseBaseEditor(SpCas9,
                                   baseEditorName="BE4max",
                                   editingStrand="original",
                                   editingWeights=ws)
metadata(BE4max)$description_base_editor <- "BE4max cytosine base editor."

}
