% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/Nuclease-class.R
\docType{class}
\name{motifs}
\alias{motifs}
\alias{motifLength}
\alias{name}
\alias{info}
\alias{weights}
\alias{name<-}
\alias{info<-}
\alias{weights<-}
\alias{cutSites}
\alias{Nuclease-class}
\alias{Nuclease}
\alias{show,Nuclease-method}
\alias{name,Nuclease-method}
\alias{name<-,Nuclease-method}
\alias{info<-,Nuclease-method}
\alias{weights,Nuclease-method}
\alias{weights<-,Nuclease-method}
\alias{motifs,Nuclease-method}
\alias{motifLength,Nuclease-method}
\alias{cutSites,Nuclease-method}
\title{An S4 class to represent a nuclease.}
\usage{
motifs(object, ...)

motifLength(object, ...)

name(object)

info(object)

weights(object, ...)

name(object) <- value

info(object) <- value

weights(object) <- value

cutSites(object, ...)

Nuclease(
  name,
  motifs = NULL,
  cutSites = NULL,
  weights = rep(1, length(motifs)),
  info = NA_character_
)

\S4method{show}{Nuclease}(object)

\S4method{name}{Nuclease}(object)

\S4method{name}{Nuclease}(object) <- value

\S4method{info}{Nuclease}(object) <- value

\S4method{weights}{Nuclease}(object, expand = FALSE)

\S4method{weights}{Nuclease}(object) <- value

\S4method{motifs}{Nuclease}(
  object,
  primary = FALSE,
  strand = c("+", "-"),
  expand = FALSE,
  as.character = FALSE
)

\S4method{motifLength}{Nuclease}(object)

\S4method{cutSites}{Nuclease}(object, strand = c("+", "-"), combine = TRUE, middle = FALSE)
}
\arguments{
\item{object}{\linkS4class{Nuclease} object.}

\item{...}{Additional arguments for class-specific methods}

\item{value}{New value to pass to the setter functions.}

\item{name}{Name of the nuclease.}

\item{motifs}{Character vector of recognition sequence motifs
written from 5' to 3' written in Rebase convention.
If the point of cleavage has been determined, the
precise site is marked with ^. Only letters in the
IUPAC code are accepted. For nucleases that cleave
away from their recognition sequence, the cleavage
sites are indicated in parentheses. See details for
more information.}

\item{cutSites}{Matrix with 2 rows (+ and - strand, respectively)
specifying the cleavage coordinates relative to the first
nucleotide of the motif sequence. Each column corresponds
to a motif specified in the \code{motifs} slot.}

\item{weights}{Optional numeric vector specifying relative weights
for the recognition motifs to specify cleavage probabilities.}

\item{info}{Optional string providing a description of the nuclease.}

\item{expand}{Should sequences be expanded to only contain ATCG nucleotides?
FALSE by default.}

\item{primary}{Should only the motif with the highest weight be returned?
FALSE by default. Only relevant if weights are stored in the 
\linkS4class{Nuclease} object.}

\item{strand}{Strand to allow reverse complementation of the motif.
"+" by default.}

\item{as.character}{Should the motif sequences be returned as a 
character vector? FALSE by default.}

\item{combine}{Should only unique values be considered?
TRUE by default.}

\item{middle}{For staggered cuts, should the middle point between
the cut on the forward strand and the cut on the reverse strand
be considered as the cut site? FALSE by default.}
}
\description{
Return motif string representations of recognition sites.

Return length of the recognition sites sequences.
}
\section{Functions}{
\itemize{
\item \code{Nuclease}: Create a \linkS4class{Nuclease} object
}}

\section{Slots}{

\describe{
\item{\code{name}}{Name of the nuclease.}

\item{\code{motifs}}{DNAStringSet of recognition sequence motifs
written from 5' to 3'.}

\item{\code{cutSites}}{Matrix with 2 rows (+ and - strand, respectively)
specifying the cleavage coordinates relative to the first
nucleotide of the motif sequence. Each column corresponds
to a motif specified in the \code{motifs} slot.}

\item{\code{weights}}{Optional numeric vector specifying relative weights
for the motifs corresponding to cleavage probabilities.}

\item{\code{info}}{Optional string providing a description of the nuclease.}
}}

\section{Constructors}{

    Use the constructor \code{link{Nuclease}} to create a Nuclease object.
}

\section{Accessors}{

\describe{
    \item{\code{name}:}{To get the name of the nuclease.} 
    \item{\code{motifs}:}{To get the recognition mofif
         nucleotide sequences.}
}
}

\examples{
EcoRI <- Nuclease("EcoRI",
                  motifs=c("G^AATTC"),
                  info="EcoRI restriction enzyme")


}
\seealso{
See the \linkS4class{CrisprNuclease} for CRISPR-specific nucleases.
}
